Adds a new health check for the cardano node that can be started by command line. The health check comamnd will return a zero code in case of a successful check and a non-zero code should the node be considered unhealthy.
The health check program checks whether the node has failed to receive
the most recent block on the chain for x minutes. The x is 10 minutes
per default. It can be changed with the command line argument
-max-time-since-last-block, which expects a Golang duration as a value.

The Dockerfile has adapted to include the health check. This Dockerfile
is also ready for cardano-node >= 1.9.0 verions.